fix storage esp breaking at high distances (#4932)

+++ src/main/java/meteordevelopment/meteorclient/utils/render/MeshVertexConsumerProvider.java
@@ -27,18 +27,18 @@ public class MeshVertexConsumerProvider implements IVertexConsumerProvider {
     }
 
     @Override
-    public void setOffset(float offsetX, float offsetY, float offsetZ) {
+    public void setOffset(int offsetX, int offsetY, int offsetZ) {
         vertexConsumer.setOffset(offsetX, offsetY, offsetZ);
     }
 
     public static class MeshVertexConsumer implements VertexConsumer {
         private final Mesh mesh;
 
-        private float offsetX, offsetY, offsetZ;
+        private int offsetX, offsetY, offsetZ;
 
-        private final float[] xs = new float[4];
-        private final float[] ys = new float[4];
-        private final float[] zs = new float[4];
+        private final double[] xs = new double[4];
+        private final double[] ys = new double[4];
+        private final double[] zs = new double[4];
         private final Color color = new Color();
 
         private int i;
@@ -47,7 +47,7 @@ public class MeshVertexConsumerProvider implements IVertexConsumerProvider {
             this.mesh = mesh;
         }
 
-        public void setOffset(float offsetX, float offsetY, float offsetZ) {
+        public void setOffset(int offsetX, int offsetY, int offsetZ) {
             this.offsetX = offsetX;
             this.offsetY = offsetY;
             this.offsetZ = offsetZ;
@@ -55,9 +55,9 @@ public class MeshVertexConsumerProvider implements IVertexConsumerProvider {
 
         @Override
         public VertexConsumer vertex(float x, float y, float z) {
-            xs[i] = offsetX + x;
-            ys[i] = offsetY + y;
-            zs[i] = offsetZ + z;
+            xs[i] = (double) offsetX + x;
+            ys[i] = (double) offsetY + y;
+            zs[i] = (double) offsetZ + z;
 
             if (++i >= 4) {
                 mesh.quad(
